# Improve performance of `runSnippetCursor`